requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<4.0
drf-spectacular>=0.27,<1.0
djangorestframework-simplejwt>=5.3,<6.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-results>=2.5,<3.0
django-celery-beat>=2.6,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-extensions>=3.2,<4.0
django-cors-headers>=4.4,<5.0
